+++ spec/unit/translators/nsidc_json_to_solr_spec.rb
@@ -229,4 +229,51 @@ describe SearchSolrTools::Translators::NsidcJsonToSolr do
       expect(facets.sort).to eql [' | TESTBIN', 'Modis Test Instrument | MODIS']
     end
   end
+
+  describe 'featured faceting' do
+    it 'translates true cumulus value to featured facet value' do
+      json_doc = { 'cumulus' => true }
+      facet = translator.translate_featured(json_doc, nil)
+      expect(facet).to include('In Earthdata Cloud')
+      expect(facet).not_to include('Show Global Only')
+    end
+
+    it 'does not translate false cumulus value to featured facet value' do
+      json_doc = { 'cumulus' => false }
+      facet = translator.translate_featured(json_doc, nil)
+      expect(facet).not_to include('In Earthdata Cloud')
+      expect(facet).not_to include('Show Global Only')
+    end
+
+    it 'does not translate missing cumulus value to featured facet value' do
+      json_doc = {}
+      facet = translator.translate_featured(json_doc, nil)
+      expect(facet).not_to include('In Earthdata Cloud')
+      expect(facet).not_to include('Show Global Only')
+    end
+
+    it 'translates global spatial coverage to featured facet value' do
+      json_doc = {}
+      spatial = [RGeo::GeoJSON.decode('type' => 'Polygon', 'coordinates' => [[[-180.0, 86.0], [180.0, 86.0], [180.0, -86.0], [-180.0, -86.0], [-180.0, 86.0]]])]
+      facet = translator.translate_featured(json_doc, spatial)
+      expect(facet).not_to include('In Earthdata Cloud')
+      expect(facet).to include('Show Global Only')
+    end
+
+    it 'does not translate non-global coverage to featured facet value' do
+      json_doc = {}
+      spatial = [RGeo::GeoJSON.decode('type' => 'Polygon', 'coordinates' => [[[-180.0, -31.23], [180.0, -31.23], [180.0, -90.0], [-180.0, -90.0], [-180.0, -31.23]]])]
+      facet = translator.translate_featured(json_doc, spatial)
+      expect(facet).not_to include('In Earthdata Cloud')
+      expect(facet).not_to include('Show Global Only')
+    end
+
+    it 'translates both true cumulus value and global spatial coverage to featured facet value' do
+      json_doc = { 'cumulus' => true }
+      spatial = [RGeo::GeoJSON.decode('type' => 'Polygon', 'coordinates' => [[[-180.0, 86.0], [180.0, 86.0], [180.0, -86.0], [-180.0, -86.0], [-180.0, 86.0]]])]
+      facet = translator.translate_featured(json_doc, spatial)
+      expect(facet).to include('In Earthdata Cloud')
+      expect(facet).to include('Show Global Only')
+    end
+  end
 end