+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+  - Updating harvesting to include cumulus value
+  - Adding a new facet that will cover both cumulus and global spatial
+    coverage.
+
 ## v6.5.1 (2023-09-26)
 
   - Updating logging default configuration